+++ library-notifications/src/test/java/com/deltadna/android/sdk/notifications/UtilsTest.kt
@@ -16,13 +16,8 @@
 
 package com.deltadna.android.sdk.notifications
 
-import android.app.ActivityManager
-import android.content.Context
-import android.content.Context.ACTIVITY_SERVICE
 import android.os.Bundle
 import com.google.common.truth.Truth.assertThat
-import com.nhaarman.mockito_kotlin.mock
-import com.nhaarman.mockito_kotlin.whenever
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
@@ -31,27 +26,6 @@ import org.robolectric.RuntimeEnvironment as RE
 
 @RunWith(RobolectricTestRunner::class)
 class UtilsTest {
-    
-    @Test
-    fun inForegroundWhenForegrounded() {
-        withProcesses(rapi("other", false), rapi()) {
-            assertThat(Utils.inForeground(this)).isTrue()
-        }
-    }
-    
-    @Test
-    fun inForegroundWhenBackgrounded() {
-        withProcesses(rapi("other"), rapi(fg = false)) {
-            assertThat(Utils.inForeground(this)).isFalse()
-        }
-    }
-    
-    @Test
-    fun inForegroundWhenOtherForegrounded() {
-        withProcesses(rapi("other")) {
-            assertThat(Utils.inForeground(this)).isFalse()
-        }
-    }
     
     @Test
     fun convertMap() {
@@ -72,24 +46,5 @@ class UtilsTest {
             assertThat(Utils.convert(this))
                     .isEqualTo("{\"1\":true,\"2\":\"string\"}")
         }
-    }
-    
-    fun withProcesses(vararg list: RAPI, block: Context.() -> Unit) {
-        val context = mock<Context>()
-        val manager = mock<ActivityManager>()
-        whenever(context.getSystemService(ACTIVITY_SERVICE)).thenReturn(manager)
-        whenever(manager.runningAppProcesses).thenReturn(list.asList())
-        whenever(context.packageName).thenReturn(RE.application.packageName)
-        
-        block.invoke(context)
-    }
-    
-    private fun rapi(
-            pkg: String = RE.application.packageName,
-            fg: Boolean = true) = with(RAPI(pkg, 0, null)) {
-        importance =
-                if (fg) RAPI.IMPORTANCE_FOREGROUND
-                else RAPI.IMPORTANCE_BACKGROUND
-        this
     }
 }