+++ library/src/test/java/com/deltadna/android/sdk/net/NetworkManagerTest.kt
@@ -30,12 +30,12 @@ import org.junit.Test
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runner.RunWith
+import org.junit.runners.JUnit4
 import org.junit.runners.model.Statement
-import org.mockito.runners.MockitoJUnitRunner
 import java.io.File
 import java.nio.charset.Charset
 
-@RunWith(MockitoJUnitRunner::class)
+@RunWith(JUnit4::class)
 class NetworkManagerTest {
     
     @Suppress("unused") // accessed by test framework