+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [4.5.0](https://github.com/deltaDNA/android-sdk/releases/tag/4.5.0) (2017-06-27)
+Added image message action event.  
+Fixed database cursor not being closed.  
+Fixed event sending on notification interactions.  
+
 ## [4.3.3](https://github.com/deltaDNA/android-sdk/releases/tag/4.3.3) (2017-06-12)
 Fixed writing to engage archive some devices.